@@ -1,8 +1,11 @@
-module.exports = {
-  stories: ['../stories/**/*.stories.js'],
-  addons: [
-    '@storybook/addon-storysource',
-    '@storybook/addon-links',
-    '@storybook/addon-essentials',
-  ],
+export default {
+  stories: ['../stories/**/*.stories.jsx'],
+  addons: ['@storybook/addon-storysource', '@storybook/addon-links', '@storybook/addon-essentials'],
+  framework: {
+    name: '@storybook/react-vite',
+    options: {}
+  },
+  docs: {
+    autodocs: true
+  }
 };
