@@ -1,3 +1,3 @@
-import MF from './components/MF';
+import MF from './components/MF.jsx';
 
 export { MF };
